# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{userId}": {
      "get": {
        "deprecated": true,
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGetDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully updated."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user by ID",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "deprecated": true,
        "operationId": "UsersController_deleteUserById",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user has been succesfully deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user by Id",
        "tags": [
          "users"
        ]
      }
    },
...
    "/profile/users/{userId}": {
      "get": {
        "deprecated": false,
        "operationId": "ProfileController_getUserById",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user profile found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSelfDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/users/{userId}/notifications": {
      "get": {
        "deprecated": false,
        "operationId": "ProfileController_getNotificationsByAccountId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDto"
                }
              }
            }
          },
          "404": {
            "description": "Notifications not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get notifications by user id",
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/users/{userId}/invitations": {
      "get": {
        "deprecated": false,
        "operationId": "ProfileController_getUserInvitations",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationWithExtraData"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get invitations received",
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/users/{userId}/invitations/{invitationId}/accept": {
      "post": {
        "deprecated": false,
        "operationId": "ProfileController_acceptInvitation",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          },
          "404": {
            "description": "Invitation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept invitation",
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/users/{userId}/proposals": {
      "get": {
        "deprecated": false,
        "operationId": "ProfileController_getUserProposals",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User proposals found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameProposalDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user proposals",
        "tags": [
          "profile"
        ]
      },
      "post": {
        "deprecated": false,
        "operationId": "ProfileController_createGameProposal",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameProposalBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game proposal created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameProposalDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid proposal data"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new game proposal",
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/users/{userId}/proposal-stats": {
      "get": {
        "deprecated": false,
        "operationId": "ProfileController_getUserProposalStats",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User proposal stats found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProposalStatsDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user proposal statistics",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "Board Vault API",
    "description": "Versioned contract snapshot for the Board Vault social group workspace API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AvatarDto": {
        "type": "object",
        "properties": {
          "backgroundColor": {
            "type": "string",
            "example": "#3B82F6"
          },
          "iconName": {
            "type": "object",
            "example": "person-fill"
          },
          "emoji": {
            "type": "object",
            "example": null
          },
          "type": {
            "type": "string",
            "example": "icon"
          },
          "initials": {
            "type": "string",
            "example": "AB"
          }
        },
        "required": [
          "backgroundColor",
          "iconName",
          "emoji",
          "type",
          "initials"
        ]
      },
      "UserGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "user@example.test"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-09-28 10:02:39"
          },
          "username": {
            "type": "string",
            "example": "sample-user"
          },
          "displayName": {
            "type": "string",
            "example": "Sample User"
          },
          "avatar": {
            "description": "The avatar of the user.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AvatarDto"
              }
            ]
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "isAdmin": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "username",
          "displayName",
          "avatar",
          "isDeleted",
          "isAdmin"
        ]
      },
      "UpdateUserBody": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "sample-user"
          },
          "displayName": {
            "type": "string",
            "example": "Sample User"
          },
          "avatar": {
            "$ref": "#/components/schemas/AvatarDto"
          }
        }
      },
...
      "SuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "description": "The success status of the operation."
          }
        },
        "required": [
          "success"
        ]
      },
...
      "Function": {
        "type": "object",
        "properties": {}
      },
...
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "accountId": {
            "type": "number",
            "example": 12345
          },
          "type": {
            "type": "string",
            "example": "expelled"
          },
          "message": {
            "type": "string",
            "example": "message"
          },
          "data": {
            "type": "object",
            "example": "{}"
          },
          "createdAt": {
            "type": "string",
            "example": "2022-03-07T16:00:00.000Z"
          },
          "isRead": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "accountId",
          "type",
          "message",
          "data",
          "createdAt",
          "isRead"
        ]
      },
...
      "UserSelfDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "user@example.test"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-09-28 10:02:39"
          },
          "username": {
            "type": "string",
            "example": "sample-user"
          },
          "displayName": {
            "type": "string",
            "example": "Sample User"
          },
          "avatar": {
            "description": "The avatar of the user.",
            "allOf": [
              {
                "$ref": "#/components/schemas/AvatarDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "username",
          "displayName",
          "avatar"
        ]
      },
      "InvitationWithExtraData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "groupId": {
            "type": "number",
            "example": 12345
          },
          "fromAccountId": {
            "type": "number",
            "example": 12345
          },
          "toAccountId": {
            "type": "number",
            "example": 12345
          },
          "sentAt": {
            "type": "string",
            "example": "2021-10-10T12:00:00Z"
          },
          "expiresAt": {
            "type": "string",
            "example": "2021-11-09T12:00:00Z",
            "description": "The invitation stops being actionable after this UTC timestamp."
          },
          "groupPersonId": {
            "type": "object",
            "example": 42,
            "nullable": true,
            "description": "Placeholder targeted by this invitation, when applicable."
          },
          "fromAccount": {
            "description": "The account data of the user who was invited.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Function"
              }
            ]
          },
          "group": {
            "description": "The group data of the group the user was invited to.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Function"
              }
            ]
          }
        },
        "required": [
          "id",
          "groupId",
          "fromAccountId",
          "toAccountId",
          "sentAt",
          "expiresAt",
          "fromAccount",
          "group"
        ]
      },
      "GameProposalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12345
          },
          "submittedBy": {
            "type": "number",
            "example": 1,
            "description": "Account ID that submitted the proposal"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "duplicate"
            ],
            "description": "Current status of the proposal"
          },
          "title": {
            "type": "string",
            "example": "Catan",
            "description": "Game title"
          },
          "imageUrl": {
            "type": "object",
            "example": "https://www.example.com/image.jpg",
            "nullable": true,
            "description": "Game image URL"
          },
          "gameAvgDuration": {
            "type": "object",
            "example": 120,
            "nullable": true,
            "description": "Average duration of the game in minutes"
          },
          "minPlayers": {
            "type": "object",
            "example": 3,
            "nullable": true,
            "description": "Minimum number of players required"
          },
          "maxPlayers": {
            "type": "object",
            "example": 4,
            "nullable": true,
            "description": "Maximum number of players allowed"
          },
          "proposedTags": {
            "type": "object",
            "example": "[\"strategy\", \"family\"]",
            "nullable": true,
            "description": "Proposed tags as JSON string"
          },
          "notes": {
            "type": "object",
            "example": "A great family strategy game",
            "nullable": true,
            "description": "User notes about the game"
          },
          "reviewedBy": {
            "type": "object",
            "example": 2,
            "nullable": true,
            "description": "Admin account ID that reviewed the proposal"
          },
          "reviewedAt": {
            "type": "object",
            "example": "2024-01-15T10:30:00Z",
            "nullable": true,
            "description": "When the proposal was reviewed"
          },
          "reviewNotes": {
            "type": "object",
            "example": "Approved - great game for our collection",
            "nullable": true,
            "description": "Admin review notes"
          },
          "createdGameId": {
            "type": "object",
            "example": 123,
            "nullable": true,
            "description": "ID of the created game if approved"
          },
          "submittedAt": {
            "type": "string",
            "example": "2024-01-10T15:45:00Z",
            "description": "When the proposal was submitted"
          }
        },
        "required": [
          "id",
          "submittedBy",
          "status",
          "title",
          "imageUrl",
          "gameAvgDuration",
          "minPlayers",
          "maxPlayers",
          "proposedTags",
          "notes",
          "reviewedBy",
          "reviewedAt",
          "reviewNotes",
          "createdGameId",
          "submittedAt"
        ]
      },
      "UserProposalStatsDto": {
        "type": "object",
        "properties": {
          "totalProposals": {
            "type": "number",
            "description": "Total number of proposals submitted",
            "example": 15
          },
          "approvedProposals": {
            "type": "number",
            "description": "Number of approved proposals",
            "example": 8
          },
          "rejectedProposals": {
            "type": "number",
            "description": "Number of rejected proposals",
            "example": 3
          },
          "duplicateProposals": {
            "type": "number",
            "description": "Number of proposals marked as duplicate",
            "example": 2
          },
          "pendingProposals": {
            "type": "number",
            "description": "Number of pending proposals",
            "example": 2
          },
          "approvalRate": {
            "type": "number",
            "description": "Approval rate percentage",
            "example": 53.33
          },
          "reputationScore": {
            "type": "number",
            "description": "User reputation score (0-100)",
            "example": 75
          }
        },
        "required": [
          "totalProposals",
          "approvedProposals",
          "rejectedProposals",
          "duplicateProposals",
          "pendingProposals",
          "approvalRate",
          "reputationScore"
        ]
      },
      "CreateGameProposalBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Catan",
            "description": "Game title"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://www.example.com/image.jpg",
            "description": "Game image URL"
          },
          "gameAvgDuration": {
            "type": "number",
            "example": 120,
            "description": "Average duration of the game in minutes"
          },
          "minPlayers": {
            "type": "number",
            "example": 3,
            "description": "Minimum number of players required"
          },
          "maxPlayers": {
            "type": "number",
            "example": 4,
            "description": "Maximum number of players allowed"
          },
          "proposedTags": {
            "type": "string",
            "example": "[\"strategy\", \"family\"]",
            "description": "Proposed tags as JSON string"
          },
          "notes": {
            "type": "string",
            "example": "A great family strategy game",
            "description": "User notes about the game"
          }
        },
        "required": [
          "title"
        ]
      },
...
    }
  }
}